Auto "--port" selection could be improved (#25)
When the `sa818.py` command is exec'd without specifying an explicit port then
a list of default device paths is scanned ("/dev/serial0", "/dev/ttyUSB0").
Unfortunately, the scan ends one once "a" device path has been found regardless
of whether the path is associated with an SA818. The `sa818.py` command, knowing
how to probe for an SA818, can continue scanning all of the known paths and in
many cases will find the device without it needing to be specified.

This change improves the default port scan.